+++ changelog.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Breaking
+- SimpleNostrTemplate: change signature of `sendPlain` method
+
 ## [0.0.2] - 2024-11-01
 
 ### Added